NextJS is for kids
