+++ packages/wallet/components/Overview.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Separator } from '@/components/ui/separator'
 import { trpc } from '@/lib/trpc'
 
@@ -12,7 +14,12 @@ export function Overview(): JSX.Element {
   const { data } = trpc.statistics.getOverview.useQuery()
 
   if (!data) {
-    return <h1>Loading...</h1>
+    return (
+    <div className="container flex bg-zinc-800 rounded-xl mt-10 justify-around">
+      <span className="basis-full text-center text-xl font-bold">
+        Overview is loading...
+      </span>
+    </div>)
   }
 
   return (
