+++ packages/wallet/lib/trpc.ts
@@ -19,5 +19,5 @@ export const trpc = createTRPCNext<RootRouter>({
       ],
     }
   },
-  ssr: true,
+  ssr: false,
 })
