+++ packages/wallet/pages/_app.tsx
@@ -1,32 +1,13 @@
 import '@/styles/globals.css'
 
-import { getChain } from '@byor/shared'
-import { EthereumClient, w3mConnectors, w3mProvider } from '@web3modal/ethereum'
-import { Web3Modal } from '@web3modal/react'
-import type { AppProps } from 'next/app'
-import { configureChains, createConfig, WagmiConfig } from 'wagmi'
+import { AppProps } from 'next/app'
 
 import { trpc } from '@/lib/trpc'
-
-const chains = [getChain()]
-const projectId = '171b81f6da969b561d747dba97534b30'
-const { publicClient } = configureChains(chains, [w3mProvider({ projectId })])
-
-const wagmiConfig = createConfig({
-  autoConnect: true,
-  connectors: w3mConnectors({ projectId, version: 1, chains }),
-  publicClient,
-})
-const ethereumClient = new EthereumClient(wagmiConfig, chains)
 
 function App({ Component, pageProps }: AppProps) {
   return (
     <>
-      <WagmiConfig config={wagmiConfig}>
         <Component {...pageProps} />
-      </WagmiConfig>
-
-      <Web3Modal projectId={projectId} ethereumClient={ethereumClient} />
     </>
   )
 }
